BD Rework
Adding new functionals to repositories
Reworking all repos
Connecting Customers Repo to the main interface

+++ DAL/Models/ReceiptsModel.cs
@@ -12,7 +12,7 @@ namespace DAL.Models
         public int? IDSUP { get; set; }
         public int? IDCOM { get; set; }
         public int? Quality { get; set; }
-        public int? Price { get; set; }
+        public decimal? Price { get; set; }
         public System.DateTime? ReceiptDate { get; set; }
     }
 }
